move table utility functions to separate file
