+++ backend/util.py
@@ -1,12 +1,10 @@
 #coding:utf-8
 import random
 from backports import csv
-from math import log, sqrt
 import sys #??????
 from tabulate import tabulate
 import io
 import openpyxl
-from copy import deepcopy, copy
 import configparser
 import codecs
 
@@ -147,33 +145,6 @@ def load_votes(votefile, consts):
 
     return parties, votes
 
-def add_totals(m):
-    """Add sums of rows and columns to a table."""
-    nm = deepcopy(m)
-    for i in range(len(m)):
-        nm[i].append(sum(m[i]))
-    totals = [sum(x) for x in zip(*nm)]
-    nm.append(totals)
-    return nm
-
-def matrix_subtraction(A, B):
-    m = len(A)
-    assert(len(B) == m)
-    if m == 0:
-        return []
-    n = len(A[0])
-    assert(all([len(A[i]) == n and len(B[i]) == n for i in range(m)]))
-    return [
-        [A[i][j] - B[i][j] for j in range(n)]
-        for i in range(m)
-    ]
-
-def find_xtd_shares(xtd_table):
-    return [[float(v)/c[-1] if c[-1]!=0 else 0 for v in c] for c in xtd_table]
-
-def find_shares(table):
-    return [[float(v)/sum(c) if sum(c)!=0 else 0 for v in c] for c in table]
-
 def print_table(data, header, labels, output, f_string=None):
     """
     Print 'data' in a table with 'header' and rows labelled with 'labels'.
@@ -259,26 +230,6 @@ def pretty_print_election(election):
     const_names.append("Total")
     xtd_results = add_totals(election.results)
     print_table(xtd_results, header, const_names, rules["output"])
-
-def entropy(votes, allocations, divisor_gen):
-    """
-    Calculate entropy of the election, taking into account votes and
-     allocations.
-     $\\sum_i \\sum_j \\sum_k \\log{v_{ij}/d_k}$, more or less.
-    """
-    assert(type(votes) == list)
-    assert(all(type(v) == list for v in votes))
-    assert(type(allocations) == list)
-    assert(all(type(a) == list for a in allocations))
-
-    e = 0
-    for c in range(len(votes)):
-        for p in range(len(votes[c])):
-            gen = divisor_gen()
-            for k in range(allocations[c][p]):
-                dk = next(gen)
-                e += log(votes[c][p]/dk)
-    return e
 
 def sim_election_rules(rs, test_method):
     """Get preset election rules for simulation from file."""
